feat: save/load system

+++ Weapon Tycoon/Assets/_Project/Scripts/Infrastructure/EventBus.cs
@@ -5,7 +5,7 @@ namespace _Project.Scripts.Infrastructure
 {
     public static class EventBus
     {
-        public static Action<int> BankIncome;
+        public static Action<long> BankIncome;
         public static Action BuyNextSpawnerPressed;
         public static Action<int> BuySpawnerUpgradeSpeedPressed;
         public static Action<int> BuySpawnerUpgradePricePressed;
