+++ Weapon Tycoon/Assets/_Project/Scripts/UI/Models/CurrencyStorage.cs
@@ -4,16 +4,16 @@ namespace _Project.Scripts.CurrencyModule.Models
 {
     public abstract class Currency
     {
-        protected int _money;
-        public int Money => _money;
+        protected long _money;
+        public long Money => _money;
 
-        public Currency(int money)
+        public Currency(long money)
         {
             _money = money;
         }
     
         [Button( "Add Money")]
-        public void AddCurrency(int amount)
+        public void AddCurrency(long amount)
         {
             _money += amount;
         }
